+++ .gitignore
@@ -158,6 +158,8 @@ celerybeat.pid
 
 # Environments
 .env
+.env.prod
+.env.dev
 .venv
 env/
 venv/
