+++ app/database/pinecone_client.py
@@ -1,15 +1,9 @@
-import os
-from dotenv import load_dotenv
 from pinecone import Pinecone
+from app.core.config import settings
 
-load_dotenv()
-
-PINECONE_API_KEY = os.getenv("PINECONE_API_KEY")
-PINECONE_INDEX = os.getenv("PINECONE_INDEX")
-
-# 새로운 Pinecone 인스턴스 방식
-pc = Pinecone(api_key=PINECONE_API_KEY)
-index = pc.Index(PINECONE_INDEX)
+# Pinecone 인스턴스 초기화
+pc = Pinecone(api_key=settings.PINECONE_API_KEY)
+index = pc.Index(settings.PINECONE_INDEX)
 
 def upsert_user_behavior_vector(user_id: str, embedding: list[float], metadata: dict = None):
     index.upsert(vectors=[
