+++ app/monitoring/metrics.py
@@ -1,0 +1,6 @@
+from prometheus_client import Counter
+
+recommendation_failures_total = Counter(
+    "recommendation_failures_total",
+    "Total number of failed recommendations"
+)
