+++ app/redis/usage.py
@@ -1,9 +1,10 @@
 import redis
 from datetime import datetime
 
+# Redis 컨테이너 도커 서비스 이름 기준 (docker-compose 기준: redis)
 r = redis.Redis(host="redis", port=6379, db=0, decode_responses=True)
 
-def get_usage_key(user_id: int, needs: str):
+def get_usage_key(user_id: int, needs: str) -> str:
     today = datetime.now().date().isoformat()
     return f"usage:{user_id}:{today}:{needs}"
 
@@ -22,7 +23,7 @@ def increment_usage(user_id: int, needs: str):
         r.expire(key, 86400)
 
 def get_all_remaining_usage(user_id: int, need_types=["contents", "creator"], max_per_day: int = 20) -> dict:
-    remaining = {}
-    for need in need_types:
-        remaining[need] = get_remaining_usage(user_id, need, max_per_day)
-    return remaining
+    return {
+        need: get_remaining_usage(user_id, need, max_per_day)
+        for need in need_types
+    }
