+++ docker-compose.dev.yml
@@ -1,0 +1,30 @@
+version: "3.9"
+
+services:
+  recommender:
+    build: .
+    container_name: trit-recommender-dev
+    ports:
+      - "8000:8000"
+    env_file:
+      - .env.dev
+    depends_on:
+      - redis
+    networks:
+      - trit-net
+    volumes:
+      - .:/app
+    command: uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload
+
+  redis:
+    image: redis:7-alpine
+    container_name: redis-dev
+    restart: unless-stopped
+    ports:
+      - "6379:6379"
+    networks:
+      - trit-net
+
+networks:
+  trit-net:
+    driver: bridge
