+++ docker-compose.prod.yml
@@ -7,13 +7,12 @@ services:
     ports:
       - "8000:8000"
     env_file:
-      - .env
+      - .env.prod
     depends_on:
       - redis
     networks:
       - trit-net
-    volumes:
-      - .:/app
+    command: uvicorn app.main:app --host 0.0.0.0 --port 8000
 
   redis:
     image: redis:7-alpine
