+++ requirements.txt
@@ -13,3 +13,4 @@ PyJWT==2.8.0
 boto3
 redis
 prometheus-fastapi-instrumentator
+pydantic-settings>=2.0.0
